+++ src/utils/SettingUtils.js
@@ -66,6 +66,10 @@ export function getSorterBackgroundColor(sorter, index, settings) {
     return index % 2 === 0 ? settings.evenColor : settings.oddColor;
 }
 
+export function getTimeDurationBackgroundColor(timeDuration, index, settings) {
+    return index % 2 === 0 ? settings.evenColor : settings.oddColor;
+}
+
 export function getLinkedObjectBackgroundColor(linkedObject, index, settings) {
     return index % 2 === 0 ? settings.evenColor : settings.oddColor;
 }